Move method for constructing message out of transport
This should enable other TransportInterface implementations
which don't need to replicate the logic for constructing a
Message instance.

Also rename MailTransport to SendGridMailTransport, so we have
space to insert alternative mail transports (like the dummy one
which initiated this work).
