+++ service-front/module/Application/src/Model/Service/ServiceAbstractFactory.php
@@ -92,7 +92,7 @@ class ServiceAbstractFactory implements AbstractFactoryInterface
             $service = new $serviceName(
                 $authenticationService,
                 $config,
-                $container->get('TwigEmailRenderer'),
+                $container->get('LocalViewRenderer'),
                 $container->get('MailTransport')
             );
         } else {
@@ -110,8 +110,10 @@ class ServiceAbstractFactory implements AbstractFactoryInterface
         }
 
         //  If required load any additional services into the resource
-        if (array_key_exists($serviceName, $this->additionalServices)
-            && is_array($this->additionalServices[$serviceName])) {
+        if (
+            array_key_exists($serviceName, $this->additionalServices)
+            && is_array($this->additionalServices[$serviceName])
+        ) {
             foreach ($this->additionalServices[$serviceName] as $setterMethod => $additionalService) {
                 if (!method_exists($service, $setterMethod)) {
                     throw new Exception(sprintf(
