+++ service-front/module/Application/src/Module.php
@@ -4,11 +4,13 @@ namespace Application;
 
 use Application\Adapter\DynamoDbKeyValueStore;
 use Application\Form\AbstractCsrfForm;
+use Application\Logging\LoggerTrait;
 use Application\Model\Service\ApiClient\Exception\ApiException;
 use Application\Model\Service\Authentication\Adapter\LpaAuthAdapter;
 use Application\Model\Service\Authentication\Identity\User as Identity;
 use Application\Model\Service\Session\PersistentSessionDetails;
 use Application\Model\Service\System\DynamoCronLock;
+use Application\View\Helper\LocalViewRenderer;
 use Alphagov\Pay\Client as GovPayClient;
 use Aws\DynamoDb\DynamoDbClient;
 use Laminas\ModuleManager\Feature\FormElementProviderInterface;
@@ -20,7 +22,6 @@ use Laminas\ServiceManager\ServiceManager;
 use Laminas\Session\Container;
 use Laminas\Stdlib\ArrayUtils;
 use Laminas\View\Model\ViewModel;
-use Opg\Lpa\Logger\LoggerTrait;
 use TheIconic\Tracking\GoogleAnalytics\Analytics;
 use Twig\Loader\FilesystemLoader;
 use Twig\Environment;
@@ -28,6 +29,8 @@ use Twig\TwigFunction;
 
 class Module implements FormElementProviderInterface
 {
+    use LoggerTrait;
+
     public function onBootstrap(MvcEvent $e)
     {
         $eventManager        = $e->getApplication()->getEventManager();
@@ -57,12 +60,14 @@ class Module implements FormElementProviderInterface
             $path = $request->getUri()->getPath();
 
             // Only bootstrap the session if it's *not* PHPUnit AND is not an excluded url.
-            if (!strstr($request->getServer('SCRIPT_NAME'), 'phpunit') &&
+            if (
+                !strstr($request->getServer('SCRIPT_NAME'), 'phpunit') &&
                 !in_array($path, [
                     // URLs excluded from creating a session
                     '/ping/elb',
                     '/ping/json',
-                ])) {
+                ])
+            ) {
                 $this->bootstrapSession($e);
                 $this->bootstrapIdentity($e, $path != '/session-state');
             }
@@ -181,14 +186,12 @@ class Module implements FormElementProviderInterface
                 },
 
                 'DynamoDbCronClient' => function (ServiceLocatorInterface $sm) {
-
                     $config = $sm->get('config')['cron']['lock']['dynamodb']['client'];
 
                     return new DynamoDbClient($config);
                 },
 
                 'DynamoCronLock' => function (ServiceLocatorInterface $sm) {
-
                     $config = $sm->get('config')['cron']['lock']['dynamodb'];
 
                     $config['keyPrefix'] = $sm->get('config')['stack']['name'];
@@ -207,6 +210,11 @@ class Module implements FormElementProviderInterface
                     ]);
                 },
 
+                'LocalViewRenderer' => function (ServiceLocatorInterface $sm) {
+                    $twigEmailRenderer = $sm->get('TwigEmailRenderer');
+                    return new LocalViewRenderer($twigEmailRenderer);
+                },
+
                 'TwigEmailRenderer' => function (ServiceLocatorInterface $sm) {
                     $loader = new FilesystemLoader('module/Application/view/email');
 
@@ -218,6 +226,7 @@ class Module implements FormElementProviderInterface
 
                     $env->registerUndefinedFunctionCallback(function ($name) use ($viewHelperManager, $renderer) {
                         if (!$viewHelperManager->has($name)) {
+                            $this->getLogger()->debug('no Twig function called ' . $name . ' for rendering emails');
                             return false;
                         }
 
@@ -242,7 +251,7 @@ class Module implements FormElementProviderInterface
     {
         return array(
             'factories' => array(
-                'StaticAssetPath' => function( $sm ){
+                'StaticAssetPath' => function ($sm) {
                     $config = $sm->get('Config');
                     return new \Application\View\Helper\StaticAssetPath($config['version']['cache']);
                 },
